pyproject.toml
[project]
name = "fkm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "msal>=1.35.1",
    "openai>=1.76.0",
    "pydantic>=2.11.3",
    "python-pptx>=1.0.2",
    "requests>=2.33.1",
]
